CHG: On <C-p> / <C-n> in the original message buffer
When the buffer is modified and a stored message is already being previewed, change the semantics of count to be interpreted relative to the currently previewed stored message.
Beforehand, one had to use increasing <C-p>, 2<C-p>, 3<C-p>, etc. to iterate through stored messages (or go to the preview window and invoke the mapping there).