BUG: Need to ingo#escape#command#mapescape() more arguments in the defined mappings
The problem occurred with a missing \ in the \| part of an a:options.ignorePattern (the | should have been escaped to <Bar>, now the \ did that job and went missing).
But messageStoreDirspec, range, and subDirForUserProvidedDirspec all could contain critical characters, too. (Only whenRangeNoMatch has a limited set of alphabetic keywords and is safe.)
Interestingly, I had already been escaping the preview command, but forgot all the other uses.