Refactoring: Invoke MessageRecall#Buffer#Recall() instead of :MessageRecall!
All the arguments are supplied to us, no need to do duplicate work (resolving (and then escaping) the message filespec). This also gets rid of the clumsy exception handling; we just need to forward the success flag.
Because :MessageRecall references past messages, but we get a:isPrevious, we need to negate the count, though.