FIX: Further recalls in the current message buffer clobber register
Because the iteration also invokes :MessageRecall!, which now considers the a:options.replacedMessageRegister.
Extend MessageRecall#Buffer#OpenNext() to just return the next replacementFilespec if a:opencmd is "return".
We can then forward that to the newly extracted s:ReplaceWithFilespec() (or bail out on already submitted error), and in there clear the a:replacedMessageRegister argument to avoid the clobbering of the register.