ENH: Enable support for excluding Git message trailers (and similar)
a:options.range can also be a List of range expressions; s:IsReplace() iterates through all until one matches and also returns that range (to reuse for replacement).
A range of a single empty line will not be deleted. s:IsReplace() also returns the rangeContents so that can be checked.

System-info: Ubuntu 20.04.6 LTS, x86_64
Platform-info: Vim 9.1.602